feat(query): implemented query handling in `RpcService`
- Refactored `shelf.rs` as `mod.rs`
- Implemented handling of `ClientQuery` inside `RpcService`
- Implemented handling of `PeerQuery` inside `RpcService`
- Moved `PeerQueryResponse` logic from `QueryService` to `RpcService`
