+++ ebi/daemon/src/services/cache.rs
@@ -1,23 +1,21 @@
 use crate::query::file_order::{FileOrder, OrderedFileSummary};
-use crate::services::peer::PeerService;
-use crate::services::rpc::{DaemonInfo};
-use crate::services::workspace::{WorkspaceService};
-use crate::shelf::shelf::{ShelfId};
+use crate::shelf::ShelfId;
 use crate::tag::{TagId, TagRef};
-use crate::workspace::{WorkspaceId};
-use std::collections::{BTreeSet};
-use std::sync::Arc;
+use crate::workspace::WorkspaceId;
+use std::collections::BTreeSet;
 
 #[derive(Clone)]
 pub struct CacheService {
-    peer_srv: PeerService,
-    workspace_srv: WorkspaceService,
-    daemon_info: Arc<DaemonInfo>,
+    //[TODO] Design Cache Structure
 }
 
 impl CacheService {
-    pub fn retrieve(&self, tag_ref: &TagRef) -> Option<BTreeSet<OrderedFileSummary>> {
-        None //[TODO] Implement caching 
+    pub fn retrieve(&self, _tag_ref: &TagRef) -> Option<BTreeSet<OrderedFileSummary>> {
+        None //[TODO] Implement cache retrieval 
+    }
+
+    pub fn store(&self, _tag_ref: &TagRef) -> Result<bool, CacheError> {
+        Ok(false) //[TODO] Implement caching 
     }
 }
 
@@ -38,5 +36,3 @@ struct HashCache {
 pub enum CacheError {
     WorkspaceNotFound,
 }
-
-pub type CacheServiceRef = Arc<CacheService>;
